+++ tests/tagging/dataset_readers/conll2003_test.py
@@ -7,10 +7,9 @@ from tests import FIXTURES_ROOT
 
 
 class TestConll2003Reader:
-    @pytest.mark.parametrize("lazy", (True, False))
     @pytest.mark.parametrize("coding_scheme", ("IOB1", "BIOUL"))
-    def test_read_from_file(self, lazy, coding_scheme):
-        conll_reader = Conll2003DatasetReader(lazy=lazy, coding_scheme=coding_scheme)
+    def test_read_from_file(self, coding_scheme):
+        conll_reader = Conll2003DatasetReader(coding_scheme=coding_scheme)
         instances = conll_reader.read(str(FIXTURES_ROOT / "tagging" / "conll2003.txt"))
         instances = ensure_list(instances)